Update Kuba parsing to reject blank values
